server/requirement.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
appdirs==1.4.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.2
async-timeout==4.0.2
attrs==19.3.0
Automat==0.8.0
backcall==0.2.0
binaryornot==0.4.4
bleach==4.1.0
blinker==1.4
blis==0.4.1
cached-property==1.5.1
cachetools==4.2.4
catalogue==1.0.0
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.10
click==8.0.3
cloud-init==22.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cookiecutter==1.7.3
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
cymem==2.0.6
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
deluge==2.0.3
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
ec2-hibinit-agent==1.0.0
entrypoints==0.3
fastapi==0.73.0
filelock==3.0.12
Flask==2.1.3
fonttools==4.28.5
frozenlist==1.2.0
fsspec==2022.1.0
fugashi==1.1.0
future==0.18.2
GeoIP==1.3.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.43.0
h11==0.9.0
hibagent==1.0.1
httplib2==0.14.0
httptools==0.1.2
hyperlink==19.0.0
idna==2.8
importlib-metadata==4.10.0
incremental==16.10.1
ipadic==1.0.0
ipykernel==6.6.1
ipython==7.31.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.1.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keyring==18.0.1
kiwisolver==1.3.2
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mecab-python3==0.7
mistune==0.8.4
mojimoji==0.0.12
more-itertools==4.2.0
multidict==5.2.0
murmurhash==1.0.6
mysql-connector-python==8.0.26
nbclient==0.5.9
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.10.4
notebook==6.4.6
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathy==0.6.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pipenv==11.9.0
plac==1.1.3
poyo==0.5.0
preshed==3.0.6
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.18.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
pydantic==1.8.2
pygame==1.9.6
Pygments==2.11.2
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.6
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
PySimpleGUI==4.60.3
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-libtorrent==1.1.13
python-multipart==0.0.5
python-slugify==6.1.1
pytorch-lightning==1.2.7
pytz==2019.3
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
regex==2021.11.10
rencode==1.0.6
requests==2.27.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.8
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.7.3
SecretStorage==2.3.1
Send2Trash==1.8.0
service-identity==18.1.0
setproctitle==1.1.10
simplejson==3.16.0
six==1.14.0
sklearn==0.0
smart-open==5.2.1
sniffio==1.2.0
sos==4.1
spacy==2.2.3
spacy-legacy==3.0.8
spacy-loggers==1.0.1
SQLAlchemy==1.4.39
srsly==1.0.5
ssh-import-id==5.10
starlette==0.17.1
SudachiDict-core==20211220
SudachiPy==0.6.2
systemd-python==234
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
texttable==1.6.2
thinc==7.3.1
threadpoolctl==3.0.0
tk==0.1.0
tokenizers==0.10.3
torch==1.7.1
torchmetrics==0.5.0
torchtext==0.4.0
torchvision==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.5.0
Twisted==18.9.0
typer==0.4.0
typing-extensions==4.0.1
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
unidic-lite==1.0.8
urllib3==1.25.8
uvicorn==0.11.3
uvloop==0.16.0
virtualenv==20.0.17
virtualenv-clone==0.3.0
wadllib==1.3.3
wasabi==0.9.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.53.0
websockets==8.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wordcloud==1.8.2.2
wsproto==0.15.0
yarl==1.7.2
youtube-dl==2020.3.24
zipp==1.0.0
zope.interface==4.7.1
